+++ tests/fixtures/templates/nocache.php
@@ -1,0 +1,1 @@
+<?php $kirby->response()->cache(false); ?><h1><?= $page->title() ?></h1>